refactor(cancellation): remove unused cancellation modules and clean up code
- Deleted atomic_labels.rs, coordinator.rs, distributed_lock.rs, recovery.rs, state_aware.rs, and mod.rs as part of a cleanup effort.
- Removed placeholder implementations and documentation to streamline the cancellation module.
- Ensured the cancellation system is more maintainable by eliminating obsolete code.

Task: 6
Status: Cleanup and refactoring complete
Priority: high
