+++ controller/src/tasks/mod.rs
@@ -11,7 +11,7 @@ use std::sync::Arc;
 use std::time::Duration;
 use tracing::{debug, error, info, instrument, Instrument};
 
-pub mod cancellation;
+pub mod cancel;
 pub mod code;
 pub mod config;
 pub mod docs;
